reface common.hpp to hetegeneous
